Isolate database data for each user account
Added new column for each table that stores the account to which the row
belongs. This is needed, for example, for identifying which "read" or
"posted" messages are presented in the tabs, since they are not the same
for each individual account. (It is not possible/straightforward to create
multiple database files for each user that logs in, since the database
is created/opened as soon as the app launches. The approach I took is
used by Google in their ContactsProvider.)
Modified content provider to add check for the current account to each
WHERE clause.